// types/MyInfo/MyInfo_EmergencyContacts.ts
export type MyInfoEmergencyContacts = {
    emergencyContacts:string;
    name: string;
    relationship: string;
    homeTelephone: string;
    mobile: string;
    workTelephone: string;
}
export type MyInfoEmergencyContacts_AttachFile = {
    attachmentEmergencyContacts: string;
    commentEmergencyContacts: string;
}